requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
Babel==2.8.0
beautifulsoup4==4.8.2
billiard==3.6.1.0
braintree==3.45.0
bs4==0.0.1
celery==4.4.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.40.0
django-ckeditor==5.8.0
django-clear-cache==0.3
django-js-asset==1.2.2
django-json-widget==1.0.0
django-jsonfield==1.4.0
django-phonenumber-field==4.0.0
django-picklefield==2.1.1
djangorestframework==3.11.0
dnspython==1.16.0
eventlet==0.25.1
fake-useragent==0.1.11
flower==0.9.2
future==0.18.2
fuzzywuzzy==0.17.0
greenlet==0.4.15
idna==2.8
isort==4.3.21
joblib==0.14.0
kombu==4.6.7
lazy-object-proxy==1.4.3
localflavor==1.8
mccabe==0.6.1
monotonic==1.5
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
pep8==1.7.1
phonenumbers==8.11.2
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-Levenshtein==0.12.0
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.3.0
scipy==1.3.2
selenium==3.141.0
six==1.11.0
soupsieve==1.9.5
sqlparse==0.2.4
tornado==4.5.2
urllib3==1.25.7
useragent==0.1.1
vine==1.3.0
virtualenv==16.7.7
wrapt==1.11.2
xlrd==1.2.0
